+++ CHANGELOG.md
@@ -3,9 +3,12 @@
 ## 0.6
 
 ### Poetry Detection
-- Update evaluation code
+- Update span-level evalaution
+  - Supports unlabeled annotations
+  - Supports sytem comparisons
   - Now include page-level F1 scores
   - Updated edge cases for precision and recall, returns 1 when denominator is 0
+- Update `get_passim_results` script so span "ref_id" field is renamed to "poem_id"
 
 ## [0.5] 2026-05-12